==> 1b_clean_data_2011.R <==
#clean 2011 data to use for sampling model (likelihood)
mydir<-'/Users/nbrodnax/Indiana/2014fall/s626_bayesian_analysis/final_project/analysis/'
setwd(mydir)

#school info
setwd("data")
Y1<-read.table("RC11p1.txt",header=F,fill=T,sep=";",stringsAsFactors=F,quote="") #school info
info<-data.frame(Y1[,1],Y1[,3],Y1[,11],Y1[,14],Y1[,20],Y1[,45],Y1[,53])
#info<-data.frame(Y1[,c(1,3,11,14,20,45,53)]) this caused type to be loaded as atomic rather than factor
labels<-c("id","name","type","black11","size11","lep11","lowinc11")
colnames(info)<-labels
info$id<-as.character(info$id)
info$type<-as.factor(info$type)
info$size11<-as.numeric(info$size11)

#achievement scores
Y2<-read.table("RC11p11.txt",header=F,fill=T,sep=";",stringsAsFactors=F,quote="") #ISAT data (no variation in naep data)
scores11<-data.frame(Y2[,c(1:5,18:21)])
labels<-c("id","rwarn","rbelow","rmeet","rexceed","mwarn","mbelow","mmeet","mexceed")
colnames(scores11)<-labels
#scores11$id<-as.character(scores11$id)
ildata<-merge(info,scores11,by="id",all=F) #merge IL data
levels(ildata$type)<-c("05","04","02","00","01") #reset the labels for school type
rm(info,scores11,Y1,Y2) #clear out stuff you don't need

#school location
common<-read.csv("common2.csv",header=T,stringsAsFactors=F,quote="") #common core data
cols<-c("school","state","charter","magnet","latitude","longitude","schoolid","schoolid2","agencyid","congress","grade4","county")
colnames(common)<-cols
common$latitude<-as.numeric(common$latitude,length=1)
common$longitude<-as.numeric(common$longitude,length=1)
common$charter<-as.factor(common$charter)
common$magnet<-as.factor(common$magnet)

#create new ids for merging il and common core data
Z<-NULL
ids<-as.vector(as.character(ildata$id))
temp<-strsplit(ids,"") #split id into separate characters
ID<-matrix(unlist(temp),ncol=15,byrow=T) #put the characters into a rowX15 matrix
for (i in 1:nrow(ID)) #create new ids
  {
  Z<-rbind(Z,paste(ID[i,1],ID[i,2],ID[i,3],ID[i,4],ID[i,5],ID[i,6],ID[i,7],ID[i,8],ildata$type[i],ID[i,13],ID[i,14],ID[i,15],sep=""))
}

ildata<-cbind(ildata,Z[,1]) #add the new ids to the dataframe
ildata[,16]<-as.character(ildata[,16]) #change new ideas from factor to character
colnames(ildata)[16]<-"schoolid" #rename new ids

data<-merge(ildata,common,by="schoolid",all=F)
write.table(data,"data.dat")

#create charter school dataset
csinfo<-read.csv("charterinfo.csv",header=F,stringsAsFactors=F,quote="",row.names=NULL)
cols<-c("key","current","open","enrolled","grades","grade4","latitude","longitude")
colnames(csinfo)<-cols
write.table(csinfo,"charter.dat")
